matrix: Demote warning about creating a new client to debug level
The latest libcmatrix will not blindly load all accounts.

However, the crash might still happen if there are multiple accounts
set up for eigenvalue
(libcmatrix checks the keyring for secrets
with a label that contains the app id)
as we currently only set the sync callback for a single client.

In any case creating a new client should not warrant raising a warning.