matrix: Save new client into DB and keyring
Needs recently introduces API namely

- cm_matrix_save_client_sync()
- cm_utils_get_homeserver_sync()

The homeserver and login ID need to be properly set before we can
successfully save the client.

`cm_utils_get_homeserver_sync()` is needed because we cannot simply
use the server part of "@user:example.com" as server deligation may be
in play and we need to query `.well-known/matrix/client` to identify
the homeserver.